Refactor remainder of commands.
